Bring id pretty printing to <fragment> support as well.
Previously only View instances in the DOM would format ids using human
readable names.  This refactors the logic into a common class and
modifies FragmentDescriptor to support this.
