+++ stetho/src/main/java/com/facebook/stetho/common/android/FragmentAccessor.java
@@ -2,6 +2,7 @@
 
 package com.facebook.stetho.common.android;
 
+import android.content.res.Resources;
 import android.view.View;
 
 import javax.annotation.Nullable;
@@ -11,6 +12,8 @@ public interface FragmentAccessor {
 
   @Nullable
   public Object getFragmentManager(Object fragment);
+
+  public Resources getResources(Object fragment);
 
   public int getId(Object fragment);
